9C-1141-4 Applications API
We refactor the api module by extracting a package `applications`,
which contains the `/application` routes.
