+++ modules/api/src/main/scala/cards/nine/api/messages/GooglePlayMessages.scala
@@ -4,44 +4,6 @@ import cards.nine.domain.application.{ Package, Widget }
 
 object GooglePlayMessages {
 
-  case class ApiCategorizedApp(
-    packageName: Package,
-    categories: List[String]
-  )
-
-  case class ApiIconApp(
-    packageName: Package,
-    title: String,
-    icon: String
-  )
-
-  case class ApiAppsInfoRequest(items: List[Package])
-
-  case class ApiAppsInfoResponse[A](errors: List[Package], items: List[A])
-
-  // ApiDetailsApp: FullCard without Screenshots
-  case class ApiDetailsApp(
-    packageName: Package,
-    title: String,
-    free: Boolean,
-    icon: String,
-    stars: Double,
-    downloads: String,
-    categories: List[String]
-  )
-
-  case class ApiSetAppInfoRequest(
-    title: String,
-    free: Boolean,
-    icon: String,
-    stars: Double,
-    downloads: String,
-    categories: List[String],
-    screenshots: List[String]
-  )
-
-  case class ApiSetAppInfoResponse()
-
   case class ApiGetRecommendationsByCategoryRequest(excludePackages: List[Package], limit: Int)
 
   case class ApiGetRecommendationsForAppsRequest(
@@ -64,8 +26,6 @@ object GooglePlayMessages {
     screenshots: List[String]
   )
 
-  case class ApiRankAppsRequest(location: Option[String], items: Map[String, List[Package]])
-
   case class ApiRankByMomentsRequest(
     location: Option[String],
     items: List[Package],
@@ -73,20 +33,8 @@ object GooglePlayMessages {
     limit: Int
   )
 
-  case class ApiRankedAppsByCategory(category: String, packages: List[Package])
-
-  case class ApiRankAppsResponse(items: List[ApiRankedAppsByCategory])
-
   case class ApiRankedWidgetsByMoment(moment: String, widgets: List[Widget])
 
   case class ApiRankWidgetsResponse(items: List[ApiRankedWidgetsByMoment])
-
-  case class ApiSearchAppsRequest(
-    query: String,
-    excludePackages: List[Package],
-    limit: Int
-  )
-
-  case class ApiSearchAppsResponse(items: List[ApiRecommendation])
 
 }
